requirements.txt
crewai
python-dotenv
chromadb
langchain-openai
tweepy
langchain-community
langchain-huggingface
sentence-transformers
pytest
pytest-mock
fastapi
